Fix #256 - Only validate the combinedFeedMessage in the CrossFeedDescriptorValidator
When this tool was originally built, some of the first rules written were for cross-referencing errors between two feeds (i.e., rules W003 and E047 in what is now CrossFeedDescriptorValidator). So, for each iteration, BackgroundTask around line 167 grabbed all the GTFS-rt messages for all GTFS-rt feeds that are being monitored simultaneously for the same GTFS data from the database and this combinedFeed was passed into the rules to be validated.

The problem with this is just that - other rules implemented in the other validator classes such as VehicleValidator were running on ALL messages from ALL GTFS-rt feeds associated with the same GTFS data. As a result, everything works fine when only one GTFS-rt feed per GTFS was running, but errors mix across multiple feeds when multiple GTFS-rt feeds were being monitored for the same GTFS data.

This patch adds another method parameter to our FeedEntityValidator interface that all rules implement specifically for the combinedFeedMessage, and passes in only the currentFeedMessage for the feedMessage parameter. CrossFeedDescriptorValidator only validates the combinedFeedMessage (not the feedMessage, which will be a subset of the combinedFeedMessage), and all other rules ignore the combinedFeedMessage and validate the feedMessage.