Add -stats command line parameter for batch processing
`-stats` *(Optional)* - If this argument is supplied (e.g., `-stats yes`), the validator will save statistics to memory for each of the validation files that are processed, and will return a list of `ValidationStatistics` objects from `BatchProcessor.processFeeds()` that can be examined to see individual iteration and rule processing times (in decimal seconds).  Default setting is false, as batch processing a very large amount of files could result in very large stats lists (one item in list per file processed).