docs: Fix docker run instructions
I'm having problems getting the container port to be accessible via the old method on Windows 10 with Docker Desktop app - I'm not sure if something changed since we originally did this. 

I suggest we change to this, which works for me (unless someone know why the previous doesn't work or how to fix it). Also note that the full name of the container needs to be used as it is in this change.